# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CanteenFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── MENU ────────────────────────────────────────────────────────────────────
...
  /seats/reserve:
    post:
      operationId: reserveSeats
      tags: [seats]
      summary: AI-powered seat reservation for a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReserveSeatsInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          description: Not enough seats available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /seats/my-reservation:
    get:
      operationId: getMyReservation
      tags: [seats]
      summary: Get current user's active reservation
      responses:
        "200":
          description: Active reservation or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationOrNull"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: cancelMyReservation
      tags: [seats]
      summary: Cancel current user's reservation
      responses:
        "204":
          description: Cancelled
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, canteen]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - name
        - email
        - role
        - createdAt
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [student, canteen]
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        seatIds:
          type: array
          items:
            type: integer
        tableNumber:
          type: string
        seatNumbers:
          type: array
          items:
            type: string
        groupSize:
          type: integer
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - seatIds
        - tableNumber
        - seatNumbers
        - groupSize
        - expiresAt
        - createdAt
    ReservationOrNull:
      oneOf:
        - $ref: "#/components/schemas/Reservation"
        - type: "null"
    ReserveSeatsInput:
      type: object
      properties:
        groupSize:
          type: integer
        preferredZone:
          type: string
      required:
        - groupSize
...